backend/no_orm/requirements-dev.txt
pytest-asyncio
httpx
mypy
black
